fix: sửa lỗi test lộ ra khi chạy lại trên môi trường thật
Chạy lại toàn bộ bộ test tích hợp trên môi trường có cơ sở dữ liệu
thật phát hiện một điểm dựng dữ liệu mẫu
trong test còn thiếu một trường bắt buộc mới thêm gần đây, khiến 2
kịch bản kiểm tra việc chặn xóa đơn hàng đã có kết quả cắt bị lỗi khi
ghi xuống cơ sở dữ liệu. Bổ sung đúng giá trị mặc định còn thiếu.

Sửa thêm 1 chỗ so sánh giá trị sai kiểu dữ liệu trong test xuất Excel
mới viết, khiến kết quả so sánh luôn báo sai dù giá trị thực tế đã
đúng - không liên quan tới logic sinh file, chỉ là cách viết điều
kiện kiểm tra.

+++ backend/src/test/java/com/slatcut/cutting/service/ExcelExportServiceTest.java
@@ -152,7 +152,9 @@ class ExcelExportServiceTest extends AbstractIntegrationTest {
             Sheet sheet = workbook.getSheet("Kết quả cắt");
             assertThat(sheet).isNotNull();
             Row header = sheet.getRow(0);
-            assertThat(header.getLastCellNum()).isEqualTo(14);
+            // getLastCellNum() trả short — AssertJ tự động box thành Short, so isEqualTo(14) autobox
+            // thành Integer thì luôn fail dù giá trị khớp (Short.equals(Integer) luôn false).
+            assertThat((int) header.getLastCellNum()).isEqualTo(14);
             assertThat(cellText(header, 0)).isEqualTo("Đợt cắt");
 
             Row dataRow = sheet.getRow(1);
